refactor(tools): reorganize utility files and update imports
- Rename batch-utils to tool-batch
- Rename readonly-constraints-validation to readonly-validation
- Extract tool specification logic to tool-spec
- Rename index to tool-policy
- Remove unused tool-call-queue
- Add pathe dependency
