+++ packages/tools/package.json
@@ -11,6 +11,7 @@
     "ai": "catalog:",
     "json-schema": "^0.4.0",
     "minimatch": "catalog:",
+    "pathe": "^2.0.3",
     "shell-quote": "^1.8.3",
     "zod": "catalog:"
   },
