+++ packages/tools/src/utils/tool-policy.ts
@@ -1,134 +1,29 @@
-import * as path from "node:path";
 import { minimatch } from "minimatch";
+import * as path from "pathe";
 import { parse } from "shell-quote";
 import type { CompiledToolPolicies } from "../types";
+import { type ToolSpecInput, getToolRules, parseToolSpec } from "./tool-spec";
 
-export type ParsedToolSpec = {
-  name: string;
-  rules: string[];
-};
+// ---------------------------------------------------------------------------
+// Policy compilation
+// ---------------------------------------------------------------------------
 
-export type ToolSpecInput =
-  | string
-  | {
-      name: string;
-      rules?: string[];
-    };
+const PathPolicyToolNames = [
+  "readFile",
+  "writeToFile",
+  "applyDiff",
+  "editNotebook",
+  "listFiles",
+  "globFiles",
+  "searchFiles",
+] as const;
 
-function sanitizeRules(rules: string[] | undefined): string[] {
-  return (rules ?? []).map((x) => x.trim()).filter((x) => x.length > 0);
-}
+type PathPolicyToolName = (typeof PathPolicyToolNames)[number];
 
-function parseToolSpecs(tools: ToolSpecInput[] | undefined): ParsedToolSpec[] {
-  return (tools ?? [])
-    .map((tool) => parseToolSpec(tool))
-    .filter((tool) => tool.name.length > 0);
-}
-
-/**
- * Parse a tool declaration such as "executeCommand(git status)".
- *
- * This helper only understands the top-level `toolName(rule)` shape used in
- * custom agent tool configuration. String declarations may contain at most one
- * rule value. If a caller needs multiple rule values, it should provide
- * multiple declarations instead of `tool(a, b)`.
- */
-export function parseToolSpec(tool: ToolSpecInput): ParsedToolSpec {
-  if (typeof tool !== "string") {
-    return {
-      name: tool.name.trim(),
-      rules: sanitizeRules(tool.rules),
-    };
-  }
-
-  const trimmed = tool.trim();
-  if (!trimmed) {
-    return {
-      name: "",
-      rules: [],
-    };
-  }
-
-  const match = trimmed.match(/^([a-zA-Z][\w-]*)\((.*)\)$/);
-  if (!match) {
-    return {
-      name: trimmed,
-      rules: [],
-    };
-  }
-
-  return {
-    name: match[1],
-    rules: parseToolRules(trimmed, match[2]),
-  };
-}
-
-function parseToolRules(tool: string, rawRule: string): string[] {
-  const rule = rawRule.trim();
-  if (!rule) {
-    return [];
-  }
-
-  if (rule.includes(",")) {
-    throw new Error(
-      `Invalid tool declaration "${tool}". Use one declaration per tool rule, for example: readFile(src/**), readFile(pochi://-/plan.md).`,
-    );
-  }
-
-  return [rule];
-}
-
-export function normalizeToolSpecs(
-  tools: ToolSpecInput[] | undefined,
-): ParsedToolSpec[] | undefined {
-  const normalized = parseToolSpecs(tools);
-
-  return normalized.length > 0 ? normalized : undefined;
-}
-
-/**
- * Collect the configured rules for a tool.
- *
- * Returning `undefined` means the tool is enabled without any policy
- * restriction. Multiple rules can still be provided by repeating tool
- * declarations or, for object-form specs, by passing multiple rules.
- */
-export function getToolRules(
-  tools: ToolSpecInput[] | undefined,
+function isPathPolicyToolName(
   toolName: string,
-): string[] | undefined {
-  let hasUnrestrictedTool = false;
-  const allowed = new Set<string>();
-
-  for (const tool of tools ?? []) {
-    const parsed = parseToolSpec(tool);
-    if (parsed.name !== toolName) {
-      continue;
-    }
-
-    const rules = parsed.rules;
-
-    if (rules.length === 0) {
-      hasUnrestrictedTool = true;
-      continue;
-    }
-
-    for (const rule of rules) {
-      allowed.add(rule);
-    }
-  }
-
-  if (hasUnrestrictedTool || allowed.size === 0) {
-    return undefined;
-  }
-
-  return [...allowed];
-}
-
-export function getAllowedToolNames(
-  tools: ToolSpecInput[] | undefined,
-): Set<string> {
-  return new Set(parseToolSpecs(tools).map((x) => x.name));
+): toolName is PathPolicyToolName {
+  return (PathPolicyToolNames as readonly string[]).includes(toolName);
 }
 
 export function compileToolPolicies(
@@ -165,24 +60,6 @@ export function compileToolPolicies(
   }
 
   return Object.keys(policies).length > 0 ? policies : undefined;
-}
-
-const PathPolicyToolNames = [
-  "readFile",
-  "writeToFile",
-  "applyDiff",
-  "editNotebook",
-  "listFiles",
-  "globFiles",
-  "searchFiles",
-] as const;
-
-type PathPolicyToolName = (typeof PathPolicyToolNames)[number];
-
-function isPathPolicyToolName(
-  toolName: string,
-): toolName is PathPolicyToolName {
-  return (PathPolicyToolNames as readonly string[]).includes(toolName);
 }
 
 function compileWebFetchDomainPolicy(tools: ToolSpecInput[] | undefined) {
@@ -242,6 +119,10 @@ function compilePathToolPolicy(
     patterns: rules,
   } as const;
 }
+
+// ---------------------------------------------------------------------------
+// Policy validation
+// ---------------------------------------------------------------------------
 
 function splitCommandSegments(command: string): string[] {
   const parsed = parse(command);
